vision.py
from PIL import ImageGrab
import os, json, base64, time

def capture_screenshot(path=None):
    path = path or os.path.join(os.path.expanduser("~"), "Documents", "ai_os_vision.png")
    img = ImageGrab.grab()
    img.save(path)
    return path

def vision_find_coord(label):
    """
    Placeholder: captures screenshot and returns None or bbox normalized [x,y,w,h].
    In production this calls Gemini vision and parses bbox.
    Here we return None so caller falls back.
    """
    p = capture_screenshot()
    return None
